✨ Adjustment of confirmation popup for registration removal

+++ Controller/Adminhtml/Author/Delete.php
@@ -12,20 +12,25 @@ class Delete extends Action
 
     public function execute()
     {
-        $id = $this->getRequest()->getParam('id');
-
         /** @var \Magento\Backend\Model\View\Result\Redirect $resultRedirect */
         $resultRedirect = $this->resultRedirectFactory->create();
 
+        $id = $this->getRequest()->getParam('id');
         if (empty($id)) {
-            $this->messageManager->addError(__('Unable to remove author.'));
-            return $resultRedirect->setPath('*/*/index');
+            $this->messageManager->addError(__('Invalid author identifier.'));
+            return $resultRedirect->setPath('*/*/edit', ['id'=>$id]);
         }
 
         try {
             /** @var \BrunoDuarte\Blog\Model\Author $author */
             $author = $this->_objectManager->create(Author::class);
+
             $author->load($id);
+            if (empty($author->getData())) {
+                $this->messageManager->addError(__('Author not found.'));
+                return $resultRedirect->setPath('*/*/edit', ['id'=>$id]);
+            }
+
             $author->delete();
 
             $this->messageManager->addSuccess(__('The Author has been removed.'));
